Pure refactor: reorganize formatter code for clarity
Rename `IndentFormatter.kt` to `Formatter.kt` and move our key `format`
function along with `KsonFormatterConfig` into there — this organizes
the code connects `Formatter.kt` and `FormatterTest.kt` in the
conventional way.

Also deprecate IndentFormatter.indent and explain what's going on in
its doc so it's clear to everyone what's going on here.
